.DS_Store file is "required" by node if found in the images directory
If a .DS_Store file is created by OS X in the images directory then natal will attempt to
require it like any other image file. This causes an error in the simulator as javascript cannot load
the file. As there isnt any reason why a .DS_Store file should be loaded I've added an automatic exclusion for this file
name in any part of the image path.

Additions that could be made:

We could expose the exclude list as a configuration in the .re-natal file so that people could see what is being excluded and add to it if needed.